+++ tests/test_linear_algebra_tools/test_LinearEquationSolver.py
@@ -1,13 +1,11 @@
 import pytest
 
 from mttools.utils.Exceptions import (
-    DimensionError,
     InconsistentWarning,
     InfiniteSolutionsWaring,
-    NoInverseWarning,
     UnderDeterminedError,
 )
-from mttools.LinearAlgebraTools.LinearAlgebraTools import solve_linear_equations
+from mttools.LinearAlgebraTools import solve_linear_equations
 
 
 class TestLinearEquationSolver:
