+++ tests/test_linear_algebra_tools/test_SquareMatrix.py
@@ -1,6 +1,6 @@
 import pytest
 
-from mttools.LinearAlgebraTools.LinearAlgebraTools import SquareMatrix
+from mttools.LinearAlgebraTools.Matrix import SquareMatrix
 from mttools.utils.Exceptions import DimensionError, NoInverseWarning
