+++ tests/test_number_theory_tools/test_NumberTools.py
@@ -1,4 +1,4 @@
-from mttools.NumberTheoryTools.NumberTools import perfect_factors
+from mttools.NumberTheoryTools import perfect_factors
 
 
 class TestPerfectFactors:
