+++ tests/test_number_theory_tools/test_PrimeTools.py
@@ -1,4 +1,4 @@
-from mttools.NumberTheoryTools.PrimeTools import (
+from mttools.NumberTheoryTools.Primes import (
     division_primality_test,
     fermat_primality_test,
     largest_prime_less_than,
